+++ src/pandemia/components/travel_model/default_travel_model_functions.c
@@ -11,7 +11,7 @@ static inline uint64_t rotl(const uint64_t x, int k) {
 }
 
 // https://prng.di.unimi.it/xoroshiro128plus.c
-uint64_t next(uint64_t * s) {
+uint64_t next(u_int64_t * s) {
 	const uint64_t s0 = s[0];
 	uint64_t s1 = s[1];
 	const uint64_t result = s0 + s1;
@@ -21,12 +21,12 @@ uint64_t next(uint64_t * s) {
 	return result;
 }
 
-int randrange(uint64_t * s, int num) {
+int randrange(u_int64_t * s, int num) {
     // Returns an int in the range [0, num)
     return next(s) % num;
 }
 
-int bernoulli(uint64_t * s, double prob_success) {
+int bernoulli(u_int64_t * s, double prob_success) {
     // Returns the value 1 if success, 0 otherwise
     int rnd = randrange(s, INT_MAX);
     int result = 0;
@@ -36,7 +36,7 @@ int bernoulli(uint64_t * s, double prob_success) {
     return result;
 }
 
-int random_choice(uint64_t * s, double * weights, double sum_of_weights) {
+int random_choice(u_int64_t * s, double * weights, double sum_of_weights) {
     // Returns an int in the range [0,l) where l is the length of weights
     int rnd = randrange(s, INT_MAX);
     int index = 0;
@@ -47,7 +47,7 @@ int random_choice(uint64_t * s, double * weights, double sum_of_weights) {
     return index;
 }
 
-void random_sample(uint64_t * s, int * sample, int n, int * population, int N) {
+void random_sample(u_int64_t * s, u_int64_t * sample, int n, u_int64_t * population, int N) {
     int t = 0; // total input records dealt with
     int m = 0; // number of items selected so far
     int u;
@@ -72,7 +72,7 @@ void close_borders
     int id,
     double scale_factor,
     double current_border_closure_multiplier,
-    int * agents_travelling_matrix,
+    u_int64_t * agents_travelling_matrix,
     const double * baseline_agents_travelling_matrix
 )
 {
@@ -96,10 +96,10 @@ void determine_travellers
     int id,
     int R, // number_of_regions
     const double * current_disease,
-    const int * current_strain,
-    int * current_region,
-    const int * agents_travelling_matrix,
-    uint64_t * random_state
+    const u_int64_t * current_strain,
+    u_int64_t * current_region,
+    const u_int64_t * agents_travelling_matrix,
+    u_int64_t * random_state
 )
 {
     int r1 = id;
@@ -111,7 +111,7 @@ void determine_travellers
     if(total_num_to_travel > 0){
 
         // Determine who is eligible to travel from this region today and count how many
-        int * agents_eligible_to_travel = (int *)malloc(sizeof(int) * N);
+        u_int64_t * agents_eligible_to_travel = (u_int64_t *)malloc(sizeof(u_int64_t) * N);
         int num_eligible_to_travel = 0;
         for(int n=0; n<N; n++){
             if(current_strain[n] == -1 && current_disease[n] < 1.0){
@@ -123,7 +123,7 @@ void determine_travellers
         // Among these eligible agents determine who actually travels today
         int num_agents_to_travel;
         num_agents_to_travel = fmin(total_num_to_travel, num_eligible_to_travel);
-        int * agents_to_travel = (int *)malloc(sizeof(int) * num_agents_to_travel);
+        u_int64_t * agents_to_travel = (u_int64_t *)malloc(sizeof(u_int64_t) * num_agents_to_travel);
         random_sample(random_state, agents_to_travel, num_agents_to_travel,
                       agents_eligible_to_travel, num_eligible_to_travel);
 
@@ -156,13 +156,13 @@ void transmission_out
     double facemask_transmission_multiplier,
     double travel_multiplier,
     double current_region_transmission_multiplier,
-    const int * current_region,
+    const u_int64_t * current_region,
     const double * current_infectiousness,
-    const int * current_strain,
-    const int * current_facemask,
+    const u_int64_t * current_strain,
+    const u_int64_t * current_facemask,
     double * sum_f_by_strain,
     double * transmission_force,
-    uint64_t * random_state
+    u_int64_t * random_state
 )
 {
     double facemask_multiplier, f;
@@ -187,15 +187,15 @@ void transmission_in
     int S, // number_of_strains
     int N, // number_of_agents
     int r1,
-    int * current_facemask,
-    int * current_region,
+    u_int64_t * current_facemask,
+    u_int64_t * current_region,
     double facemask_transmission_multiplier,
     double * sum_f_by_strain,
     double * current_sigma_immunity_failure,
-    int * infection_event,
+    u_int64_t * infection_event,
     double * transmission_force,
     double * mutation_matrix,
-    uint64_t * random_state
+    u_int64_t * random_state
 )
 {
     double * weights = (double *)malloc(sizeof(double) * S);
